FIX: issue 4378: Generic JDBC specific number->float conversion
git-svn-id: https://svn.javlin.eu/svn/cloveretl.opensource/trunk@9238 a09ad3ba-1a0f-0410-b1b9-c67202f10d70

+++ cloveretl.connection/src/org/jetel/connection/jdbc/specific/impl/AbstractJdbcSpecific.java
@@ -124,7 +124,7 @@ abstract public class AbstractJdbcSpecific implements JdbcSpecific {
 		case DataFieldMetadata.INTEGER_FIELD:
 			return Types.INTEGER;
 		case DataFieldMetadata.NUMERIC_FIELD:
-			return Types.NUMERIC;
+			return Types.FLOAT;
 		case DataFieldMetadata.STRING_FIELD:
 			return field.isFixed() ? Types.CHAR : Types.VARCHAR;
 		case DataFieldMetadata.DATE_FIELD:
